ref: RenderCacheService aus Timeline.razor ausgelagert. fix: Detail Panel durch auslagerung und CSS refactoring zerstört. Repariert und Design dem allgemeinen Design angepasst.

+++ Components/Models/Timeline/Anomaly.cs
@@ -8,7 +8,7 @@
 
         public string Url { get; set; } = "";
 
-        public double Position { get; set; }
+        public int Position { get; set; }
 
         public DomainEntry LinkedDomain { get; set; } = null!;
